% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{CPBcorr}
\alias{CPBcorr}
\title{Codon pair bias correlation test}
\usage{
CPBcorr(refOne, refTwo)
}
\arguments{
\item{refOne}{First CPB reference table}

\item{refTwo}{Second CPB reference table}
}
\value{
Output from \code{cor.test}, and a scatter plot.
}
\description{
Spearman rank correlation test for comparing two CPB reference tables
}
\details{
Uses \code{cor.test} R base function with the spearman method. Depending on the CPB's, it might be useful to standardize the data first.
}
\examples{
CPBcorr(Homo.sapiens, Aedes.aegypti)
}

