% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpscalc.R
\name{CPScalc}
\alias{CPScalc}
\title{Codon pair score calculator}
\usage{
CPScalc(sequence, reference, start = 1, end = NULL, draw = TRUE,
  windowSize = NULL, silent = FALSE)
}
\arguments{
\item{sequence}{Sequences can be input directly as a character string or as the file path to a fasta file.}

\item{reference}{Reference CPB table. See \code{\link{CPBtable}}.}

\item{start}{Nucleotide position in the sequence marking the region of the sequence to use for calculation.}

\item{end}{Nucleotide position in the sequence marking the region of the sequence to use for calculation.}

\item{draw}{If TRUE, a line plot showing the local CPS along the length of the sequence is output to the graphics device.}

\item{windowSize}{CPS line plots are smoothed by locally weighted polynomial regression where \code{windowSize} designates the number of nucleotides over which individual codon pair scores are smoothed. If NULL, smoothing spans 7.5\% of the sequence length.}

\item{silent}{If TRUE the progress bar is suppressed.}
}
\value{
An invisible list is returned:
\item{averageCPS}{Average of the individual codon pair scores of the input sequence.}
\item{totalCPS}{Numeric vector containing the individual codon pair scores along the sequence length.}
}
\description{
Calculate the codon pair score of a sequence relative to a reference codon pair bias.
}
\examples{
fastaLocation <- system.file('tbevns5.fasta', package = 'CPBias')
tbev <- importFasta(fastaLocation)[[1]]
tbevCPS <- CPScalc(tbev, Homo.sapiens)
tbevCPS[[1]]

# Can plot a smoothed line plot with CPScalc output as it is done in
# CPSplot()
CPSx <- 1:length(tbevCPS[[2]])
smoothCPS <- loess(tbevCPS[[2]] ~ CPSx, span = (.05), degree = 2)
lineCPS <- predict(smoothCPS)
plot(lineCPS, type='l')
}

