% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{CPSplot}
\alias{CPSplot}
\title{Codon pair score line plot}
\usage{
CPSplot(seqOne, refOne, seqTwo = NULL, refTwo = NULL, start = 1,
  end = NULL, windowSize = NULL)
}
\arguments{
\item{seqOne}{Sequences can be input directly as a character string or as the file path to a fasta file. All sequences must be in the correct reading frame, stop codons or codons not defined in the translation table are not allowed and will generate an error.}

\item{refOne}{CPB reference table. See \code{\link{CPBtable}}.}

\item{seqTwo}{An alternate codon pair recoding of seqOne. seqTwo only accepts sequences with identical length to seqOne.}

\item{refTwo}{A second CPB reference. The sequence(s) will be plotted relative to both references.}

\item{start}{Nucleotide position in the sequence.}

\item{end}{Nucleotide position in the sequence.}

\item{windowSize}{Designates the number of nucleotides over which smoothing occurs. If NULL windowSize is 7.5\% of the sequence length.}
}
\value{
Line plot in the graphics device.
}
\description{
Local codon pair scores plotted as a smooth line along the length of a nucleotide sequence.
}
\details{
The individual codon pair scores across a gene sequence are plotted as a smooth curve by locally weighted quadratic regression (\code{loess}). Span is defined as the window size divided by total sequence length in nucleotides. Two different CPS encodings of a sequence can be compared with up to two reference CPB's.
}
\examples{
fastaLocation <- system.file('tbevns5.fasta', package = 'CPBias')
tbev <- importFasta(fastaLocation)[[1]]

humCPS <- CPScalc(tbev, Homo.sapiens, silent=TRUE, draw=FALSE)[[1]]
tbevhmax <- CPSdesign.dual(tbev, Aedes.aegypti, Homo.sapiens, humCPS, 'max',
draw=FALSE, silent=TRUE)
CPSplot(tbev, Homo.sapiens, tbevhmax[[9]], Homo.sapiens)
}

