% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{importFasta}
\alias{importFasta}
\title{Import nucleotide sequence from fasta file}
\usage{
importFasta(fastaLocation, sepSequences = FALSE)
}
\arguments{
\item{fastaLocation}{Location of fasta file.}

\item{sepSequences}{If FALSE fasta files containing multiple sequences are concatenated into one string, otherwise a vector of strings is returned.}
}
\value{
\item{seqs}{Sequence(s).}
\item{seqNames}{Names of all sequences.}
}
\description{
A fasta file containing single or multiple sequences is imported as a lowercase character string, or a vector of multiple character strings.
}
\details{
Comments marked by `;' are ignored. If importing sequences for use in \code{\link{CPBtable}}, make sure to set \code{sepSequences} to TRUE otherwise sequences not divisible by three might shift the read frame.
}
\examples{
# Tick borne encephalitis virus NS5 coding region
fastaLocation <- system.file('tbevns5.fasta', package = 'CPBias')
tbev <- importFasta(fastaLocation)
tbev[[2]]

# Randomly selected protein coding sequences from the human
# CCDS data set.
fastaLocation <- system.file('ccds.fasta', package = 'CPBias')
ccds <- importFasta(fastaLocation, sepSequences = TRUE)
ccds[[2]]
}

