% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datarox.R
\name{listCPB}
\alias{listCPB}
\title{Codon pair bias reference tables}
\format{\describe{
\item{Species}{ Species scientific name.}
\item{EMBL.ID}{ EMBL WGS project ID number.}
\item{Taxonomy}{ Species lineage.}
\item{Number.of.Sequences}{ Number of sequences used in CPB calculation.}
\item{CPB.reference}{ Name of R object containing the CPB reference table.}
}}
\source{
\url{http://www.ebi.ac.uk/ena}
}
\description{
Table of species specific CPB reference tables and information about the sequences used to create them.
}
\details{
Sequences were collected for those organisms which have whole genome sequence data available in the EMBL database (ENA Release 121). Only protein coding sequences were used, mitochondrion and chloroplast sequences were omitted.
}
\note{
Only the Homo.sapiens and Aedes.aegypti CPB reference table data sets are included in this package, the rest of the references listed can be downloaded at \url{https://github.com/zayets} as rda files and loaded directly in R.
}
\examples{
# write to CSV
data(listCPB)
write.csv(listCPB, 'listCPB.csv', row.names=FALSE)
}

