% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datarox.R
\name{REseqs}
\alias{REseqs}
\title{Restriction enzyme table}
\format{\describe{
\item{Enzyme}{ Restriction enzyme.}
\item{Prototype}{ Prototype enzyme for the recognition sequence.}
\item{Organism}{ Organism in which enzyme was isolated.}
\item{Recognition.sequence}{ 5' to 3' restriction enzyme recognition sequence, with cleavage marked where applicable.}
\item{Regex.sequence}{ 5' to 3' recognition sequence represented as a regular expression.}
}}
\source{
\url{http://rebase.neb.com}
}
\description{
Restriction enzyme recognition sequences in the form of R regular expressions.
}

