% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datarox.R
\name{standardTranslation}
\alias{standardTranslation}
\title{Standard translation table}
\format{\describe{
\item{codon}{ Individual codons.}
\item{amino}{ One letter identifiers for amino acids or `STOP' for stop codons.}
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi}
}
\description{
Standard code translation table as a two column data frame containing all unique coding and non-coding codons in column one and a single letter identifier for the encoded amino acid in column two.
}

